perf: use integral offset over pointer
This change modifies the internal offset for the Bump
allocator to be of type std::size_t instead of std::byte*.
The primary motivation for this change was to allow one
single atomic lookup instead of two. Previosly, to check
if a request_size is within range of the current headroom,
the pointer of the tail AND the offset pointer were both
atomically fetched. This was unnecessary since the block sizes
are statically defined constants. We can use that information
to check headroom instead of touching the tail pointer.