fix: add mutex to Page
Page allocator was thread-unsafe, causing the downstream
Bump tests to be flaky. This occured because the Page
allocator uses a simple array to store all page requests.
Occasionally, two or more threads will allocate a page
and have it assigned to the same index within the array.
Thus, one of these allocations would be lost. When it wastime for memory release, one of
the lost pages will trigger an error, since the release function
could not find an entry in its table. No good.

Added a simple mutex as a proof of concept and now tests
pass consistently. However, this is not ideal as the mutex
is coarse-grained and we want a lock-free algorithm.