feat: introduce BlockArray class
This class will store page span entries for Page provider
classes. It allows for quick insertions, O(1), and but
slow removals in worst case O(N). Note, re-arranging
elements in random-access removal is fast as we can
swap with the last element in array. This is so because
we do not care about the order of the elements